Cambios de imagen

+++ src/app/interceptors/api-interceptor.ts
@@ -1,13 +1,18 @@
 import { HttpInterceptorFn, HttpErrorResponse } from '@angular/common/http';
-import { inject } from '@angular/core';
+import { inject, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 import { Router } from '@angular/router';
 import { catchError, throwError } from 'rxjs';
 import { environment } from '../../environments/environment';
 
 export const apiInterceptor: HttpInterceptorFn = (req, next) => {
 
-  const token = localStorage.getItem('token');
   const router = inject(Router);
+  const platformId = inject(PLATFORM_ID);
+  const isBrowser = isPlatformBrowser(platformId);
+
+  // ✅ Solo accede a localStorage si estamos en el navegador
+  const token = isBrowser ? localStorage.getItem('token') : null;
 
   let request = req;
 
@@ -31,10 +36,10 @@ export const apiInterceptor: HttpInterceptorFn = (req, next) => {
 
   return next(request).pipe(
     catchError((error: HttpErrorResponse) => {
-      if (error.status === 401) {
+      if (error.status === 401 && isBrowser) {
         localStorage.clear();
         router.navigate(['/login'], {
-          queryParams: { sessionExpired: 'true' } // 👈 le avisas al login
+          queryParams: { sessionExpired: 'true' }
         });
       }
       return throwError(() => error);
